refactor: complete structural review — callDepth underflow detection and cleanup
Replace the silent clamp guard in PopContinuation with an unconditional
decrement and panic on negative, making callDepth underflow immediately
detectable now that the field is int rather than uint64. Fix stale uint64(0)
comparison in pool_test.go. Expand remaining single-line function definition
(integerPassthrough in extensions/math). Mark all structural review items
done in STRUCTURAL_REVIEW.md.

+++ machine/pool_test.go
@@ -243,7 +243,7 @@ func TestAcquireContinuation_ReturnsZeroedFrame(t *testing.T) {
 	qt.Assert(t, cont.singleValue, qt.IsNil)
 	qt.Assert(t, cont.multiValues, qt.IsNil)
 	qt.Assert(t, cont.pc, qt.Equals, 0)
-	qt.Assert(t, cont.callDepth, qt.Equals, uint64(0))
+	qt.Assert(t, cont.callDepth, qt.Equals, 0)
 	qt.Assert(t, cont.promptHandler, qt.IsNil)
 	qt.Assert(t, cont.promptTag, qt.IsNil)
 	qt.Assert(t, cont.shared, qt.IsFalse)
